refactor(matrix): subject is matrix inner service
