+++ internal/matrix/endpoints/create_subject.go
@@ -11,8 +11,7 @@ import (
 	kithttp "github.com/go-kit/kit/transport/http"
 	"github.com/google/uuid"
 
-	domain2 "github.com/sumelms/microservice-course/internal/matrix/domain"
-	"github.com/sumelms/microservice-course/internal/subject/domain"
+	domain "github.com/sumelms/microservice-course/internal/matrix/domain"
 	"github.com/sumelms/microservice-course/pkg/validator"
 )
 
@@ -57,7 +56,7 @@ func makeCreateSubjectEndpoint(s domain.ServiceInterface) endpoint.Endpoint {
 			return nil, err
 		}
 
-		c := domain2.Subject{}
+		c := domain.Subject{}
 		data, _ := json.Marshal(req)
 		if err := json.Unmarshal(data, &c); err != nil {
 			return nil, err
