+++ internal/matrix/endpoints/delete_subject.go
@@ -10,7 +10,7 @@ import (
 	"github.com/google/uuid"
 	"github.com/gorilla/mux"
 
-	"github.com/sumelms/microservice-course/internal/subject/domain"
+	"github.com/sumelms/microservice-course/internal/matrix/domain"
 )
 
 type deleteSubjectRequest struct {
