+++ internal/matrix/endpoints/list_subject.go
@@ -7,7 +7,7 @@ import (
 	"github.com/go-kit/kit/endpoint"
 	kithttp "github.com/go-kit/kit/transport/http"
 
-	"github.com/sumelms/microservice-course/internal/subject/domain"
+	"github.com/sumelms/microservice-course/internal/matrix/domain"
 )
 
 type listSubjectResponse struct {
