+++ internal/matrix/endpoints/update_subject.go
@@ -12,8 +12,7 @@ import (
 	"github.com/google/uuid"
 	"github.com/gorilla/mux"
 
-	domain2 "github.com/sumelms/microservice-course/internal/matrix/domain"
-	"github.com/sumelms/microservice-course/internal/subject/domain"
+	"github.com/sumelms/microservice-course/internal/matrix/domain"
 	"github.com/sumelms/microservice-course/pkg/validator"
 )
 
@@ -59,7 +58,7 @@ func makeUpdateSubjectEndpoint(s domain.ServiceInterface) endpoint.Endpoint {
 			return nil, err
 		}
 
-		c := domain2.Subject{}
+		c := domain.Subject{}
 		data, _ := json.Marshal(req)
 		if err := json.Unmarshal(data, &c); err != nil {
 			return nil, err
