+++ internal/matrix/transport/http.go
@@ -32,9 +32,15 @@ func NewHTTPHandler(r *mux.Router, s domain.ServiceInterface, logger log.Logger)
 	r.Handle("/matrices/{uuid}", updateMatrixHandler).Methods(http.MethodPut)
 	r.Handle("/matrices/{uuid}", deleteMatrixHandler).Methods(http.MethodDelete)
 
-	addSubjectHandler := endpoints.NewAddSubjectHandler(s, opts...)
-	removeSubjectHandler := endpoints.NewRemoveSubjectHandler(s, opts...)
+	listSubjectHandler := endpoints.NewListSubjectHandler(s, opts...)
+	createSubjectHandler := endpoints.NewCreateSubjectHandler(s, opts...)
+	findSubjectHandler := endpoints.NewFindSubjectHandler(s, opts...)
+	updateSubjectHandler := endpoints.NewUpdateSubjectHandler(s, opts...)
+	deleteSubjectHandler := endpoints.NewDeleteSubjectHandler(s, opts...)
 
-	r.Handle("/matrices/subject", addSubjectHandler).Methods(http.MethodPost)
-	r.Handle("/matrices/subject", removeSubjectHandler).Methods(http.MethodDelete)
+	r.Handle("/subjects", createSubjectHandler).Methods(http.MethodPost)
+	r.Handle("/subjects", listSubjectHandler).Methods(http.MethodGet)
+	r.Handle("/subjects/{uuid}", findSubjectHandler).Methods(http.MethodGet)
+	r.Handle("/subjects/{uuid}", updateSubjectHandler).Methods(http.MethodPut)
+	r.Handle("/subjects/{uuid}", deleteSubjectHandler).Methods(http.MethodDelete)
 }
